Create module "vufs" and move DTO`s from common-utils
